Initial Fork, Decoupling of Player Movement and Input
This iteration decouples the movement of the player and the input to their own classes to simplify the code and make it more modular.

Some changes were also made to some of the other class files, such as fixing indentation, the names of variables and functions so that the names make sense, as well as removing redundant code.